lesson 7 resolve - verify

+++ src/services/auth.service.ts
@@ -70,31 +70,32 @@ class AuthService {
 
     const hashedPassword = await passwordService.hash(dto.password);
 
-    // const users = await userRepository.getAll();
-    //
-    // await Promise.all(
-    //   users.map(async (user) => {
-    //     await emailService.sendMail(user.email, EEmailAction.WELCOME, {
-    //       name: user.name,
-    //     });
-    //   }),
-    // );
-
-    // await emailService.sendMail(dto.email);
-
-    //todo
-    // await emailService.sendMail(dto.email, EEmailAction.WELCOME, {
-    //   name: dto.name,
-    // });
-    //todo for joi to work otherwise uncomment
-
-    //if an array of emails -
-    // await emailService.sendMail(["malaniako@gmail.com", "katya@gmail.com"], EEmailAction.WELCOME);
-
-    return await userRepository.create({
+    const user = await userRepository.create({
       ...dto,
       password: hashedPassword,
     });
+
+    const actionToken = tokenService.createActionToken(
+      {
+        userId: user._id,
+        role: ERole.USER,
+      },
+      EActionTokenType.ACTIVATE,
+    );
+
+    await Promise.all([
+      tokenRepository.createActionToken({
+        actionToken,
+        _userId: user._id,
+        tokenType: EActionTokenType.ACTIVATE,
+      }),
+      emailService.sendMail(dto.email, EEmailAction.WELCOME, {
+        name: dto.name,
+        actionToken,
+      }),
+    ]);
+
+    return user;
   }
   public async signIn(dto: ILogin): Promise<ITokensPair> {
     const user = await userRepository.getOneByParams({ email: dto.email });
@@ -107,6 +108,9 @@ class AuthService {
 
     if (!isMatch) {
       throw new ApiError("Not valid email or password", 401);
+    }
+    if (user.isVerified) {
+      throw new ApiError("Verify your account", 403);
     }
 
     const jwtTokens = tokenService.generateTokenPair(
@@ -182,6 +186,24 @@ class AuthService {
       await userRepository.updateById(payload.userId, {
         password: newHashedPassword,
       }),
+      tokenRepository.deleteActionTokenByParams({ actionToken }),
+    ]);
+  }
+
+  public async verify(actionToken: string) {
+    const payload = tokenService.checkActionToken(
+      actionToken,
+      EActionTokenType.ACTIVATE,
+    );
+    const entity = await tokenRepository.getActionTokenByParams({
+      actionToken,
+    });
+    if (!entity) {
+      throw new ApiError("Token not valid", 400);
+    }
+
+    await Promise.all([
+      await userRepository.updateById(payload.userId, { isVerified: true }),
       tokenRepository.deleteActionTokenByParams({ actionToken }),
     ]);
   }
